Product + product detail modul (admin), varian module, color module

+++ src/main/java/bca/Hapepedia/entity/Varian.java
@@ -13,7 +13,7 @@ import javax.persistence.Table;
 public class Varian {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(length = 20, nullable = false, unique = true)
     private String name;
     @ManyToOne
@@ -22,16 +22,16 @@ public class Varian {
     public Varian() {
     }
 
-    public Varian(int id, String name) {
+    public Varian(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
